refactor: extract Simba ERP menus component
